Kind of working better. Several range finding changes.
Moved IIR leakage out of timers and put a timer token in timers instead.
Changed the IIR to float from double which is presumably the same thing.
Brought the IRR value in as an extern to avoid function calls and passing by stack, etc.

Don't understand floats very well, need to try to do this with integer math.