More work towards stabilizing the range detection so most of the changes were to range.c. Believe the old floating point version may have had some memory leak.
Changed the IIR filter from using floats to using 32 bit integers.
The range sensor (in mm usually about 110-120 as an 8 bit number) is scaled by 65536.
The iir coefficients are now parts of an 8 bit EEPROM value (iir_value) so they are iir)value and 256-iir_value.
The old floating point coefficients we about 0.88 and 0.12 but should be set in EEPROM to be about 235 which means 0.918 and 0.082.
Likewise the leakage factor has been converted to a scaled version of itself (now 65536 times higher it was 0.00024 now is just 16).

Finally the "steady foot counter was removed" as kind of silly.
If it's still twitchy will try upping the filter time constant some more.