Attempting to factor ALS reading out of timers caused the lockups to come back.
Evidence is that there was some (interrupt driven) state machine interaction with the taking of ALS readings which are admittedly blocking.
The blocking ALS reading is called from two places.
Added interrupt guards around each call.
Successfully removed calls to trigger the ALS reading directly from within the timers module.
The sensor module already had the start_ALS command anyway so it was being done twice which was flaky to begin with.

We'll see how this goes...