Giant rip and redo. Substantially a copy of the rework done with the VL5130 sensor on the Awesome(R)Faucet project. Just ported the sensor calls over to this project and got it working again.
Future work will probably be done in this project since this is the one I have clean hardware for and it seems to be no worse than the installed code base so far.
The improved SCPI command set and floating point math will make tweaking this a lot easier if there are false trips.